requirements.txt
pandas==2.0.3
numpy==1.24.3
aiohttp==3.8.5
redis==4.6.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
sqlalchemy==2.0.20
